# admin/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PLC4X Manager API",
    "description": "REST API for managing Apache PLC4X OPC-UA Server. Configure devices, read/write PLC tags, manage users, certificates, backups, and server lifecycle.",
    "version": "1.0.0",
    "contact": {
      "name": "Silvio Garbes",
      "url": "https://github.com/silviogarbes/plc4x-manager"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {"url": "http://localhost:8080", "description": "Local (HTTP)"},
    {"url": "https://localhost:8443", "description": "Local (HTTPS)"}
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth/login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for machine-to-machine access (set API_KEY in .env)"
      }
    },
    "schemas": {
      "Device": {
        "type": "object",
        "required": ["name", "connectionString"],
        "properties": {
          "name": {"type": "string", "example": "ControlLogix1"},
          "connectionString": {"type": "string", "example": "logix://192.168.1.50"},
          "enabled": {"type": "boolean", "default": true},
          "tags": {"type": "array", "items": {"$ref": "#/components/schemas/Tag"}},
          "calculatedTags": {"type": "array", "items": {"$ref": "#/components/schemas/CalculatedTag"}}
        }
      },
      "Tag": {
        "type": "object",
        "required": ["alias", "address"],
        "properties": {
          "alias": {"type": "string", "example": "Temperature"},
          "address": {"type": "string", "example": "%DB1:0:REAL"}
        }
      },
      "CalculatedTag": {
        "type": "object",
        "required": ["alias", "formula"],
        "properties": {
          "alias": {"type": "string", "example": "TempFahrenheit"},
          "formula": {"type": "string", "example": "Temperature * 9 / 5 + 32"}
        }
      },
...
    }
  },
  "security": [{"BearerAuth": []}, {"ApiKeyAuth": []}],
...
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": ["Configuration"],
        "summary": "Get full configuration",
        "responses": {"200": {"description": "Full config as JSON"}}
      },
      "put": {
        "tags": ["Configuration"],
        "summary": "Save full configuration (replaces ALL config)",
        "description": "WARNING: This replaces the entire configuration including all devices. Use GET first, modify, then PUT back.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Config saved"}, "400": {"description": "Invalid config"}}
      }
    },
    "/api/config/server": {
      "put": {
        "tags": ["Configuration"],
        "summary": "Update server settings only",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string", "example": "Plc4xOPCUAServer"},
                  "tcpPort": {"type": "integer", "example": 12687},
                  "disableInsecureEndpoint": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Server config updated"}}
      }
    },
    "/api/devices": {
      "get": {
        "tags": ["Devices"],
        "summary": "List all devices",
        "responses": {"200": {"description": "Array of devices", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Device"}}}}}}
      },
      "post": {
        "tags": ["Devices"],
        "summary": "Add a device",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Device"}}}},
        "responses": {"201": {"description": "Device added"}, "400": {"description": "Invalid input or tag address"}, "409": {"description": "Device already exists"}}
      }
    },
...
  }
}